Add Current Scratch Code
I have some scratch code that I have used to test the viability of using
typescript to create a plugin architecture for a traditional todo app. I
want to show that you do not need to couple to a particular javascript
web framework to create a todo app. Ideally, I would like to create a
plugin architecture that exposes interfaces that any framework can use
to power the todo app.