Add knockout plugin POC
I created a set of classes that can power a UI in a framework agnostic
way. I use knockout as an example of how to implement these classes with
a particular implementation.

I still have to do a lot of work to complete it. However, I have
implemented enough of the functionality to justify a commit.