Add backbone plugin
I wanted to provide a backbone implementation of my application.

I downloaded backbone and all its dependencies, and I tied them together
with the TodoListFacade in index.html.

This plugin implements the add item functionality only. I need to add
support to update the status of an item, and remove an item.

I will do that some time later.