Add a .gitignore
My machine shows me the VI metadata files when I try to commit or add files. I
don't want git to show me these files; so, I have created a .gitignore file. I
decided to commit this to the git repo because I have multiple machines that I
work on and this happens quite frequently on all of them.